// frontend/src/app/models/payment.model.ts
export class Payment {
    share_id: number;
    card_number: string;
    exp_year: string;
    exp_month: string;
    cvc: string;
    name: string;
    last_name: string;
    email: string;
    doc_number: string;
    city: string;
    address: string;
    phone: string;
    cell_phone: string;
  }